+++ checkpoint/orbax/checkpoint/experimental/v1/_src/testing/compatibility/load_pytree_compatibility_test.py
@@ -26,6 +26,7 @@ import orbax.checkpoint.experimental.v1 as ocp
 from orbax.checkpoint.experimental.v1._src.context import options as options_lib
 from orbax.checkpoint.experimental.v1._src.handlers import registration
 from orbax.checkpoint.experimental.v1._src.layout import checkpoint_layout as checkpoint_layout_lib
+from orbax.checkpoint.experimental.v1._src.testing.compatibility import test_utils as compatibility_test_utils
 
 
 CheckpointLayoutEnum = options_lib.CheckpointLayout
@@ -48,40 +49,6 @@ class LoadPytreeCompatibilityTest(parameterized.TestCase):
     self.abstract_state = jax.tree.map(
         lambda x: jax.ShapeDtypeStruct(x.shape, x.dtype, sharding=sharding),
         self.expected_state
-    )
-
-  def get_checkpoint_path(
-      self,
-      version: str,
-      metadata_present: bool,
-      is_direct_checkpoint: bool,
-      is_pytree: bool,
-  ) -> epath.Path | None:
-    """Returns the path to the checkpoint for each combination of parameters."""
-    if version == 'v1' and is_direct_checkpoint:
-      return None  # V1 does not support direct checkpoints.
-
-    version_dir = f'{version}_checkpoints'
-    type_dir = (
-        'direct_checkpoint' if is_direct_checkpoint else 'composite_checkpoint'
-    )
-    metadata_dir = (
-        'checkpoint_metadata_present'
-        if metadata_present
-        else 'checkpoint_metadata_missing'
-    )
-    pytree_dir = (
-        'pytree_checkpointable_has_metadata'
-        if is_pytree
-        else 'pytree_checkpointable_missing_metadata'
-    )
-
-    return (
-        self.base_dir
-        / version_dir
-        / type_dir
-        / metadata_dir
-        / pytree_dir
     )
 
   def setup_registry(
@@ -226,7 +193,7 @@ class LoadPytreeCompatibilityTest(parameterized.TestCase):
       handler_registered: bool,
       pytree_registered: bool,
   ) -> None:
-    path = self.get_checkpoint_path(
+    path = compatibility_test_utils.get_checkpoint_path(
         version, metadata_present, is_direct_checkpoint, is_pytree
     )
     if path is None or not path.exists():